requirements.txt
addict==2.4.0
anyio==4.12.0
appnope==0.1.4
asttokens==3.0.1
certifi==2025.11.12
click==8.3.1
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
executing==2.2.1
filelock==3.20.2
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
humanfriendly==10.0
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
jedi==0.19.2
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
markdown-it-py==4.0.0
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mmengine==0.10.7
mpmath==1.3.0
narwhals==2.14.0
nest-asyncio==1.6.0
numpy==2.2.6
onnxruntime==1.23.2
opencv-python==4.12.0.88
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
pillow_heif==1.1.1
platformdirs==4.5.1
plotly==6.5.0
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
pyzmq==27.1.0
rich==14.2.0
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
termcolor==3.3.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typer-slim==0.21.0
typing_extensions==4.15.0
wcwidth==0.2.14
widgetsnbextension==4.0.15
yapf==0.43.0
